Rename AeIdentityMain to AeIdentityLight
