+++ src/components/aeIdentityLight/aeIdentityLight.scss
@@ -1,6 +1,6 @@
 @import './../variables';
 
-.ae-identity-main._invert {
+.ae-identity-light._invert {
   color: $white;
 }
 
@@ -75,7 +75,7 @@
   padding-right: 3px;
 }
 
-.ae-identity-main .avatar {
+.ae-identity-light .avatar {
   border-width: 1px;
   width: 23px;
   height: 23px;
