+++ src/components/aeIdentityLight/index.js
@@ -1,0 +1,5 @@
+import aeIdentityLight from './aeIdentityLight.vue'
+
+export default function install (Vue) {
+  Vue.component('ae-identity-light', aeIdentityLight)
+}
