--- src/components/aeIdentityMain/index.js
@@ -1,5 +1,0 @@
-import aeIdentityMain from './aeIdentityMain.vue'
-
-export default function install (Vue) {
-  Vue.component('ae-identity-main', aeIdentityMain)
-}
